+++ lib/shadcn_ui.ex
@@ -25,6 +25,7 @@ defmodule ShadcnUI do
     ShadcnUI.Components.Forms.Help,
     ShadcnUI.Components.Forms.Input,
     ShadcnUI.Components.Forms.Label,
+    ShadcnUI.Components.Forms.RadioGroup,
     ShadcnUI.Components.Forms.Textarea
   ]